==> Codes/3c_build_cnn.ipynb <==
# https://machinelearningmastery.com/pytorch-tutorial-develop-deep-learning-models/
# https://colab.research.google.com/github/rpi-techfundamentals/website_spring_2020/blob/master/content/notebooks/20-deep-learning1/06-regression-bh-pytorch.ipynb#scrollTo=xD9PhAU7hoqT
#!pip install torchvision
import numpy as np
import pandas as pd
import pyreadr
import scipy
import torch
import torch.nn as nn
import torch.nn.functional as F
from sklearn.utils import shuffle
from torch.autograd import Variable


class Net(torch.nn.Module):
    def __init__(self, cols, size_hidden, n_output):
        super(Net, self).__init__()
        self.hidden = torch.nn.Linear(cols, size_hidden)   # hidden layer
        self.predict = torch.nn.Linear(size_hidden, n_output)   # output layer

    def forward(self, x):
        x = F.relu(self.hidden(x))      # activation function for hidden layer
        x = self.predict(x)             # linear output
        return x

def build_cnn(TrainSet_eonr2_df, policy):
    y_train = TrainSet_eonr2_df['eonr']
    X_train = TrainSet_eonr2_df.drop('eonr', axis=1)
    #Define training hyperprameters.
    batch_size = 50
    num_epochs = 200
    learning_rate = 0.01
    size_hidden= 100
    
    #Calculate some other hyperparameters based on data.  
    batch_no = len(X_train) // batch_size  #batches
    cols=X_train.shape[1] #Number of columns in input matrix
    n_output=1
    #Create the model
    device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
    # Assume that we are on a CUDA machine, then this should print a CUDA device:
    print("Executing the model on :",device)
    
    net = Net(cols, size_hidden, n_output)
    #Adam is a specific flavor of gradient decent which is typically better
    optimizer = torch.optim.Adam(net.parameters(), lr=learning_rate)
    #optimizer = torch.optim.SGD(net.parameters(), lr=0.2)
    criterion = torch.nn.MSELoss(size_average=False)  # this is for regression mean squared loss
    X_train=X_train.values
    y_train=y_train.values
    running_loss = 0.0
    for epoch in range(num_epochs):
        #Shuffle just mixes up the dataset between epocs
        X_train, y_train = shuffle(X_train, y_train)
        # Mini batch learning
        for i in range(batch_no):
            start = i * batch_size
            end = start + batch_size
            inputs = Variable(torch.FloatTensor(X_train[start:end]))
            labels = Variable(torch.FloatTensor(y_train[start:end]))
            # zero the parameter gradients
            optimizer.zero_grad()

            # forward + backward + optimize
            outputs = net(inputs)
            #print("outputs",outputs)
            #print("outputs",outputs,outputs.shape,"labels",labels, labels.shape)
            loss = criterion(outputs, torch.unsqueeze(labels,dim=1))
            loss.backward()
            optimizer.step()

            # print statistics
            running_loss += loss.item()

        #print('Epoch {}'.format(epoch+1), "loss: ",running_loss)
        running_loss = 0.0
        path = '/home/germanm2/n_policy_box/Data/files_rds/cnn_models/'+ policy + '.pth'
        torch.save(net.state_dict(), path)
        return(net)

# Build one CNN
TrainSet_eonr2_df = pyreadr.read_r("/home/germanm2/n_policy_box/Data/files_rds/TrainSet_eonr2.rds")[None] # also works for RData
policy = 'ratio_5'
net_returned = build_cnn(TrainSet_eonr2_df, policy)
net_returned

def predict_cnn(prediction_set_aggregated_dt, policy, net):
    #Initialize the eonr model:
    cols=21 #Number of columns in input matrix        
    n_output=1        
    size_hidden= 100  
    
    #Load the cnn model
    #path = '/home/germanm2/n_policy_box/Data/files_rds/cnn_models/'+ policy + '.pth'
    # net = torch.load(path)
    #net.load_state_dict(torch.load(path))
    #net.eval()
    
    #Get X data ready
    X_pred=prediction_set_aggregated_dt.values
    X = Variable(torch.FloatTensor(X_pred)) 
  
    #Make predictions
    y_pred = net(X) #This outputs the value for regression

    y_pred=y_pred.data[:,0].numpy()

    prediction_set_aggregated_dt['eonr_pred'] = y_pred
    return(prediction_set_aggregated_dt)

policy = 'ratio_5'
#Load data
prediction_set_aggregated_dt = pyreadr.read_r("/home/germanm2/n_policy_box/Data/files_rds/prediction_set_aggregated_dt.rds")[None] # also works for RData
prediction_set_aggregated_dt = prediction_set_aggregated_dt[['rain_30', 'rain_60','rain_90', 't_max_30', 't_max_60', 't_max_90', 't_min_30', 't_min_60', 't_min_90', 'Y_prev',
    'Y_corn_lt_avg', 'day_sow', 'day_v5', 'lai_v5', 'whc', 'oc_20cm_v5', 'sw_dep_v5', 'n_0_60cm_v5', 'surfaceom_wt_v5', 'sand_40cm', 'clay_40cm']]

# 

predict_cnn(prediction_set_aggregated_dt, policy, net_returned)

policy = 'ratio_5'
path = '/home/germanm2/n_policy_box/Data/files_rds/cnn_models/'+ policy + '.pth'
net2 = Net(21, 100, 1)
net2.state_dict()


net2.load_state_dict(torch.load(path))
net2.eval()
net2.state_dict()

net_returned.state_dict()

#Get X data ready
X_pred=prediction_set_aggregated_dt.values
X = Variable(torch.FloatTensor(X_pred)) 
X
#Make predictions
# y_pred = net(X) #This outputs the value for regression
net_returned(X)

 Initialize the maturity prediction model:
        model = MaturityPrediction()
        
        for model_code, model_field in field_codes.items():
            # Define the location of the saved weigths:
            model_file_name = f'../data/{model_code}_{fold}_rnd_1_noaug.pth'

            # Load model parameters:
            model.load_state_dict(torch.load(model_file_name))

            # Load the model into the GPU:
            model = model.to(device)
            model.eval()

model = Net()
model.load_state_dict(torch.load(path))

X_pred = prediction_set_aggregated_df[['rain_30', 'rain_60','rain_90', 't_max_30', 't_max_60', 't_max_90', 't_min_30', 't_min_60', 't_min_90', 'Y_prev',
 'Y_corn_lt_avg', 'day_sow', 'day_v5', 'lai_v5', 'whc', 'oc_20cm_v5', 'sw_dep_v5', 'n_0_60cm_v5', 'surfaceom_wt_v5', 'sand_40cm', 'clay_40cm']]
X_pred=X_pred.values

X = Variable(torch.FloatTensor(X_pred)) 
y_pred = net(X) #This outputs the value for regression
y_pred=y_pred.data[:,0].numpy()

prediction_set_aggregated_df['eonr_pred'] = y_pred


#import pandas as pd
X = Variable(torch.FloatTensor(X_train)) 
result = net(X)
# pred=result.data[:,0].numpy()
# print(len(pred),len(y_train))
# r2_score(pred,y_train)

X_pred = prediction_set_aggregated_df[['rain_30', 'rain_60','rain_90', 't_max_30', 't_max_60', 't_max_90', 't_min_30', 't_min_60', 't_min_90', 'Y_prev',
 'Y_corn_lt_avg', 'day_sow', 'day_v5', 'lai_v5', 'whc', 'oc_20cm_v5', 'sw_dep_v5', 'n_0_60cm_v5', 'surfaceom_wt_v5', 'sand_40cm', 'clay_40cm']]
X_pred=X_pred.values
X_pred

# MAKE PREDICTIONS
X = Variable(torch.FloatTensor(X_pred)) 
y_pred = net(X) #This outputs the value for regression
y_pred=y_pred.data[:,0].numpy()
y_pred

prediction_set_aggregated_df['eonr_pred'] = y_pred
prediction_set_aggregated_df.head

# now let's write a Rds
pyreadr.write_rds("/home/germanm2/n_policy_box/Data/files_rds/prediction_set_aggregated_cnn_df.rds", prediction_set_aggregated_df)
